refactor bios and tweets to use same rich text rendering
