+++ frontend/src/components/BioContent.jsx
@@ -1,0 +1,13 @@
+import RichText from "./RichText";
+
+function BioContent({ bio }) {
+  return (
+    bio && (
+      <div className="user-bio text-white text-xl">
+        <RichText text={bio.content} />
+      </div>
+    )
+  );
+}
+
+export default BioContent;
