+++ frontend/src/components/TweetContent.jsx
@@ -1,56 +1,15 @@
 import "./Tweet.css";
+import RichText from "./RichText";
 import PropTypes from "prop-types";
 
 function TweetContent({ tweet }) {
-  const renderText = (text) => {
-    const hashtagRegex = /(#\w+)/g; // match on # followed by 1 or more letters, numbers, or underscores
-    const usernameRegex = /(@[\w]+)/g; // match on @ followed by 1 or more letters, numbers, or underscores
-    const urlRegex = /(https?:\/\/[^\s]+)/g; // match on http(s) followed by :// and non-whitespace characters
-
-    const parts = text.split(/(#\w+|@\w+|https?:\/\/[^\s]+)/);
-
-    return parts.map((part, index) => {
-      if (part.match(urlRegex)) {
-        const displayText =
-          part.replace(/^https?:\/\/(www\.)?/, "").slice(0, 33) + // update bot logic if you change URL display len
-          (part.length > 33 ? "..." : "");
-        return (
-          <a
-            key={index}
-            href={part}
-            className="text-[#008080] underline hover:text-[#00abab]"
-          >
-            {displayText}
-          </a>
-        );
-      } else if (part.match(hashtagRegex)) {
-        const hashtag = part.slice(1);
-        return (
-          <a
-            key={index}
-            href={`/hashtag/${hashtag}`}
-            className="text-[#008080] underline hover:text-[#00abab]"
-          >
-            {part}
-          </a>
-        );
-      } else if (part.match(usernameRegex)) {
-        const username = part.slice(1);
-        return (
-          <a
-            key={index}
-            href={`/profile/${username}`}
-            className="text-[#008080] underline hover:text-[#00abab]"
-          >
-            {part}
-          </a>
-        );
-      }
-      return part;
-    });
-  };
-
-  return tweet && <div className="tweet-content">{renderText(tweet.text)}</div>;
+  return (
+    tweet && (
+      <div className="tweet-content">
+        <RichText text={tweet.text} />
+      </div>
+    )
+  );
 }
 
 TweetContent.propTypes = {
